# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bet Room Service API
  version: 1.0.0
  description: |
    The bet room service: a strategy-agnostic room container plus the `duel`
    strategy, serving the routes the `@sport-widgets/bet-room` widget calls.

    Prose, provenance and the reasoning behind every decision here:
    `bet-room-api.md` in this directory. Read it before changing anything — the
    wire-format constraints in its section 1 are not expressible in OpenAPI and
    each one breaks the widget SILENTLY.

    Three of them in short:

      * Success bodies are NOT enveloped. The body IS the object.
      * Error bodies use `message`, not `error`.
      * All timestamps are epoch MILLISECONDS as JSON numbers, never RFC3339.

    `POST /rooms/{roomId}/read` is a POST on purpose: the invite code is a join
    secret and a query string is the one place it must never appear.

    NOT IMPLEMENTED, and documented rather than hidden:

      * Server-side eligibility re-checks on create and join (two-way, prematch,
        live prices) need a sportsbook feed this service does not have. The
        underround guard IS enforced, on both paths, because it is pure
        arithmetic over stored odds. "The event has started" is NOT enforced.
      * Settlement. Nothing writes `status: settled | void` or
        `payload.winnerParticipantId`, so a filled duel stays filled forever.
servers:
  - url: http://localhost:5000
    description: Local docker compose
  - url: http://localhost:5001
...
security:
  - bearerAuth: []
paths:
  /rooms:
    post:
      tags: [rooms]
      summary: Create a room
      description: |
        The creator's bet is placed FIRST, in the host bet slip, and the room is
        opened from it — so `betRef` is always in hand and is required.

        The server derives `capacity` from `strategy` (2 for `duel`) and never reads
        it from the request. It clamps `expiresAt` to
        `min(request.expiresAt, createdAt + inviteWindow, eventStart)`; the client
        sends its own computation, but the client is not a trust boundary, so the
        server may only ever shorten it.

        `inviteCode` is issued here and must NOT be derivable from the room id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: Unrecognised strategy, or malformed body. No room is stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "409":
          description: Daily duel limit reached, or stake above the per-duel maximum.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: |
            Reserved for "the event has already started", which is NOT enforced:
            it needs a sportsbook feed this service does not have. Documented so
            the status is not reused for something else.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/Error"

    get:
      tags: [rooms]
      summary: List open rooms on an event
      description: |
        Returns rooms on the event that are neither filled nor expired, excluding
        the caller's own rooms and excluding rematch rooms (which are joinable by
        one named opponent only and must never be publicly listed).

        An event with no open rooms returns `[]`, never a 404.

        DECIDED: this response includes `inviteCode` for the rooms it lists —
        listing a duel on an event IS the offer to anyone viewing it. See
        `bet-room-api.md` section 2.5 for the consequence, which must not be
        "hardened" away later.
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: string
          example: evt-ucl-2026-09-15
        - name: status
          in: query
          required: true
          description: Only `open` is called by the client today.
          schema:
            type: string
            enum: [open]
      security:
        - bearerAuth: []
        - {}
      description: |
        No credential is required, so an event's open duels are visible to a
        spectator who has not signed in. A signed-in caller additionally gets
        `inviteCode` and `inviteUrl` on each room, which is what reserving a seat
        needs; both are omitted for an anonymous one, who could not act on them
        anyway and would otherwise be able to harvest every open code on an event
        in one request. The caller's own rooms and all rematch rooms are excluded.
      responses:
        "200":
          description: Open rooms on the event, possibly empty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "500":
          $ref: "#/components/responses/Error"
...
  /rooms/{roomId}/read:
    post:
      tags: [rooms]
      summary: Read room state
      description: |
        POST, not GET: the invite code is a join secret and must never reach a
        query string. A participant reads without a code; a code holder supplies one.

        No credential is required: a room is observable, so a widget can render a
        duel to a spectator who has not signed in. What the credential changes is
        WHAT comes back. `inviteCode` and `inviteUrl` are returned only to a
        participant or to a caller who presented the correct code — the url embeds
        the code as `betRoomInvite`, so the two are disclosed together or not at
        all. Every other reader gets room state with both fields absent, which
        keeps a scraped room id from being worth as much as a code. Taking a seat
        still requires a token.

        The response must set `viewerParticipantId` for the caller — a reader
        cannot otherwise identify their own side, and matching on the display label
        is unsafe because labels are not unique.

        This is the polled route and dominates request volume: ~5s while waiting,
        ~30s while matched, nothing in terminal states.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                inviteCode:
                  type: string
                  example: d4k9xq
      responses:
        "200":
          description: Current room state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required on every route. The token travels only in the Authorization
        header — never a query string, never a body, both of which are logged and
        cached where a header is not.

        The participant payload carries no user identifier at all, so this token is
        the service's ONLY caller identity. Four things depend on resolving a
        subject from it: `viewerParticipantId`, the creator-cannot-take-a-second-seat
        rule, excluding the caller's own rooms from the list, and the rematch
        opponent restriction.
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
      example: room-4f21
...
  responses:
    Error:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorised:
      description: Missing or invalid bearer credential. Nothing is created or changed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The room or hold referenced does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      description: |
        NOTE the key is `message`, NOT `error`. The client reads `body.message` and
        falls back to the HTTP status text. The live service emits `{"error": ...}`
        today, which would drop every stated reason. `code` is additive — for logs
        and future client branching; the widget reads only `message`.
      properties:
        message:
          type: string
          example: This duel has already been taken
        code:
          type: string
          example: roomFilled
    Timestamp:
      type: integer
      format: int64
      description: |
        Epoch MILLISECONDS. Not RFC3339 — Go's time.Time marshals to RFC3339 by
        default and the client parses this as a number.
      example: 1757596400000
    BetRef:
      type: object
      required: [id, number]
      description: |
        Reference to a participant's own ordinary sportsbook single bet — the only
        link the room keeps to money. The bet's stake, odds and payout are
        deliberately NOT carried: the room is a coordinator, never a wallet.
      properties:
        id:
          type: string
          example: bet-8812
        number:
          type: integer
          example: 12345
    ParticipantPayload:
      type: object
      required: [label, initials, balances]
      description: |
        What the client sends about the acting user on create and on join.

        The label arrives ALREADY DERIVED ("Maksym K." plus initials "MK"). The
        client never sends raw name fields, never an email address, never an account
        number and never a credential — the bearer token identifies the caller, so
        the payload needs no identifier at all. The service must not start asking
        for more: a service that is not a profile store never holds a legal name.

        `label` and `initials` are untrusted display text originating from another
        player's account.
      properties:
        label:
          type: string
          example: Maksym K.
        initials:
          type: string
          example: MK
        balances:
          type: object
          additionalProperties: true
          description: |
            Carried so a later balance-aware strategy needs no second lookup.
            OPAQUE: stored and returned unchanged, never computed with, never used
            to authorise anything.
          example:
            eur: 75.5
    Participant:
      allOf:
        - $ref: "#/components/schemas/ParticipantPayload"
        - type: object
          required: [id, betRef]
          description: |
            A CONFIRMED participant. `betRef` is required: a participant exists only
            once their commitment is confirmed, so a hold that lapsed or was
            released cannot be represented as one.

            The caller's token subject is stored alongside these fields and is NEVER
            returned.
          properties:
            id:
              type: string
              description: Opaque id issued by the room service; scoped to this room.
              example: p-1
            betRef:
              $ref: "#/components/schemas/BetRef"
...
    RoomStatus:
      type: string
      description: |
        ONE explicit value, never a set of independent booleans — mutually exclusive
        states must not be representable at once. This replaces the current
        `isFilled` / `isStarted` pair.

        A room counts as filled in `filled`, `settled` and `void`, and NEVER in
        `expired` — that is where a room which did not fill ends up. A room that
        filled before its expiry instant is unaffected by that instant passing.

        Whether the EVENT has started is deliberately not room state: that is a
        question for the sportsbook feed, and duplicating it here would let the two
        disagree.
      enum: [open, reserved, filled, expired, settled, void]
    Room:
      type: object
      required:
        [id, eventId, strategy, status, capacity, participants, createdAt, expiresAt, payload]
      properties:
        id:
          type: string
          example: room-4f21
        eventId:
          type: string
          description: Sportsbook event the room concerns. Rooms are discoverable by it.
          example: evt-ucl-2026-09-15
        strategy:
          $ref: "#/components/schemas/Strategy"
        status:
          $ref: "#/components/schemas/RoomStatus"
        capacity:
          type: integer
          description: |
            Declared by the strategy, not by the request. 2 for `duel`.
          example: 2
        participants:
          type: array
          description: Confirmed participants only. A live hold is not listed here.
          items:
            $ref: "#/components/schemas/Participant"
        viewerParticipantId:
          type: string
          description: |
            Which participant the reading caller is. Omitted when the reader holds
            only the invite code and has not taken a seat. The client sends no
            identifier of its own, so it cannot work this out any other way — and
            matching on the display label is unsafe because labels are not unique.
          example: p-1
        inviteCode:
          type: string
          description: |
            The join secret. Must NOT be derivable from `id`. Omitted unless the
            caller may hold it — see `GET /rooms` for the one case where it is
            returned to a non-participant, and why.
          example: d4k9xq
        inviteUrl:
          type: string
          description: |
            A HOST event-page URL carrying `betRoomId` and `betRoomInvite`, NOT a
            URL on this service. Both parameters are required: the client has no
            resolve-by-code-alone route, so a half-formed link is treated as no
            invite at all.

            This supersedes the earlier ask for a `GET /join/{id}` route here, and
            resolves the standing gotcha that the boilerplate's `roomUrl` 404'd
            in a browser — the link was never this service's to serve.

            `PUBLIC_BASE_URL` is this service's own base and is the WRONG base for
            it; a separate host-event-URL template belongs in config.
          example: https://sportsbook.example/event/evt-ucl-2026-09-15?betRoomId=room-4f21&betRoomInvite=d4k9xq
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        expiresAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
            - description: |
                When an unfilled room stops accepting seats. For a duel this is
                `min(createdAt + inviteWindow, eventStart)`, which is what makes
                "cannot be accepted after kick-off" a property of the data rather
                than a check someone might forget.

                Reads and the list query must evaluate this against now, not rely
                on a sweeper having run — otherwise a room reads as open and the
                widget offers a seat that cannot be taken.
        payload:
          $ref: "#/components/schemas/DuelPayload"
        rematchOfRoomId:
          type: string
          description: |
            Set on a rematch: the room this one descends from. The service resolves
            the single participant of that room who may take the open seat.
          example: room-3a08
    Strategy:
      type: string
      description: |
        Chosen at creation. Determines capacity, eligibility rules, entry arithmetic
        and settlement interpretation. Lifecycle behaviour is identical across
        strategies. An unrecognised value is a 400 and stores no room.
      enum: [duel]
...
    CreateRoomRequest:
      type: object
      required: [eventId, strategy, participant, betRef, expiresAt, payload]
      description: |
        Unlike ConfirmSeatRequest and RematchRequest, this body needs no phasing:
        the shipped client already sends every field here, including the fully
        populated DuelPayload with `creatorSide.placement`. So everything below can
        be required from day one, and `DuelPayload`'s own required list is the
        end-state shape rather than a target.
      properties:
        eventId:
          type: string
          example: evt-ucl-2026-09-15
        strategy:
          $ref: "#/components/schemas/Strategy"
        participant:
          $ref: "#/components/schemas/ParticipantPayload"
        betRef:
          allOf:
            - $ref: "#/components/schemas/BetRef"
            - description: |
                The creator's ALREADY-PLACED bet. Required, not optional: creation
                places the bet before the room exists, so the reference is always in
                hand. Omitting it leaves a room holding a participant whose only
                link to money is empty — and that link is the whole basis of the
                room being a coordinator rather than a wallet.
        expiresAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
            - description: |
                The client's own `min(createdAt + inviteWindow, eventStart)`. The
                server CLAMPS it — the client is not a trust boundary — and may only
                ever shorten it.
        payload:
          $ref: "#/components/schemas/DuelPayload"
...
    SideOffer:
      type: object
      required: [outcomeId, odd]
      properties:
        outcomeId:
          type: string
          example: over
        odd:
          type: integer
          description: |
            RAW INTEGER odds: 182 means 1.82. Keep them integers server-side — the
            underround guard is exact integer arithmetic and introduces no floating
            point at all.
          example: 182
    SidePlacement:
      type: object
      required: [participantId, stake, lineItemId, dataVersion]
      description: |
        A side once its bet exists. `lineItemId` and `dataVersion` are the two
        feed-sourced fields a bet request requires and the market feed does not
        carry; they come from a GetOutcomesV2 subscription. They are recorded as
        placed and are never read back to place a LATER bet — a stored `dataVersion`
        is stale by the time the opposite side is taken.
      properties:
        participantId:
          type: string
          example: p-1
        stake:
          type: number
          description: Stake actually placed on this side, in account currency, 2dp.
          example: 10.00
        lineItemId:
          type: string
          example: li-12
        dataVersion:
          type: integer
          example: 7
    EntryFigures:
      type: object
      required: [payout, entry, pot]
      description: |
        payout = round(creatorStake * creatorOdd, 2)
        entry  = floor(payout / opponentOdd, 2)     <- floor, from the ROUNDED payout
        pot    = creatorStake + entry

        The order is load-bearing: flooring the exact product instead of the rounded
        payout gives 0.61 where the correct path gives 0.60. The entry is floored so
        the opponent's own return never exceeds the stated winner's take.

        Use integer minor units or a decimal type, never float64 — 8.87 * 100 is
        886.9999999999999 and floors to 8.86.

        The gap between `pot` and `payout` is the book's existing margin on the two
        prices. It is not a fee, and the room never takes it.

        Provisional while the room is unfilled: the joiner's entry is recomputed
        against live prices at acceptance.
      properties:
        payout:
          type: number
          description: What the winner takes.
          example: 18.20
        entry:
          type: number
          description: What the joiner puts in, always rounded DOWN.
          example: 8.87
        pot:
          type: number
          example: 18.87
    DuelPayload:
      type: object
      required: [marketId, marketItemId, creatorSide, opponentSide, figures]
      description: |
        The duel strategy's payload, carried by the room generically — but the
        service WRITES into it on confirm and on settlement, so for `duel` it cannot
        be a blind blob.

        Decode it as a raw message and validate per-strategy: `decodeJSON` uses
        DisallowUnknownFields, which applies to nested structs too, so a typed
        struct here turns any future client field into a 400 on the one route whose
        job is to carry strategy data opaquely.

        The selection is the TRIPLE marketId + marketItemId + a side's outcomeId.
        The market and item sit at payload level, which structurally guarantees both
        sides are on the same line: a market holds many items, so a market id alone
        cannot tell Over 2.5 from Over 3.5.
      properties:
        marketId:
          type: string
          example: total_goals
        marketItemId:
          type: string
          example: total_goals_2.5
        creatorSide:
          allOf:
            - $ref: "#/components/schemas/SideOffer"
            - type: object
              required: [placement]
              description: |
                ALWAYS placed: a duel room is created from a bet that already
                exists, because the creator's stake is only known once the host bet
                slip has placed it.
              properties:
                placement:
                  $ref: "#/components/schemas/SidePlacement"
        opponentSide:
          allOf:
            - $ref: "#/components/schemas/SideOffer"
            - type: object
              description: |
                Offered from creation, placed only on acceptance — the opposite
                ordering, because a joiner's entry is committed on their behalf at a
                fixed amount and must never exist without a seat.
              properties:
                placement:
                  $ref: "#/components/schemas/SidePlacement"
        figures:
          $ref: "#/components/schemas/EntryFigures"
        winnerParticipantId:
          type: string
          description: |
            The participant whose side settled as a win. Absent while settlement is
            pending, and absent on a void.

            DEFERRED: nothing in the client can write this. It is set by a
            settlement worker resolving each participant's `betRef` against the
            sportsbook's settlement — a dependency this service does not have.
            Until that ships, every filled duel stays `filled` forever and the
            settled, closed and rematch presentations are unreachable.

            The duel never PAYS the winner: the sportsbook's settlement of that
            participant's own bet is the payment.
          example: p-1